# docs/swagger.yaml
basePath: /api
definitions:
...
  models.AccountType:
    enum:
    - full
    - temporary
    type: string
    x-enum-comments:
      AccountTypeFull: Volledige account met app toegang
      AccountTypeTemporary: Tijdelijke registratie alleen voor event
    x-enum-descriptions:
    - Volledige account met app toegang
    - Tijdelijke registratie alleen voor event
    x-enum-varnames:
    - AccountTypeFull
    - AccountTypeTemporary
...
  models.EventConfig:
    additionalProperties: true
    type: object
...
  models.EventResponse:
    properties:
      description:
        type: string
      end_time:
        type: string
      event_config:
        $ref: '#/definitions/models.EventConfig'
      geofences:
        items:
          $ref: '#/definitions/models.Geofence'
        type: array
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      start_time:
        description: ISO 8601 format
        type: string
      status:
        description: 'V27: Direct database column'
        type: string
      status_description:
        description: 'V27: Lookup table description'
        type: string
    type: object
...
  models.Geofence:
    properties:
      lat:
        description: Latitude
        type: number
      long:
        description: Longitude
        type: number
      name:
        type: string
      radius:
        description: Radius in meters
        type: number
      type:
        description: '"start", "checkpoint", "finish"'
        type: string
    type: object
...
  models.PublicRegistrationRequest:
    properties:
      afstand:
        enum:
        - 2.5 KM
        - 6 KM
        - 10 KM
        - 15 KM
        type: string
      bijzonderheden:
        description: Verplicht als Ondersteuning = Ja of Anders
        type: string
      email:
        type: string
      event_id:
        description: Event context
        type: string
      heeft_vervoer:
        description: 'V37: Transport vraag'
        type: boolean
      naam:
        description: Persoonlijke gegevens
        type: string
      ondersteuning:
        enum:
        - Ja
        - Nee
        - Anders
        type: string
      rol:
        description: Event keuzes
        enum:
        - Deelnemer
        - Begeleider
        - Vrijwilliger
        type: string
      telefoon:
        description: Optioneel, maar verplicht voor Begeleider/Vrijwilliger
        type: string
      terms:
        description: Voorwaarden
        type: boolean
      test_mode:
        description: Test mode
        type: boolean
      wachtwoord:
        description: Verplicht als WantAccount = true
        type: string
      want_account:
        description: 'Account keuze - V30: Nieuw duaal systeem'
        type: boolean
    required:
    - afstand
    - email
    - heeft_vervoer
    - naam
    - ondersteuning
    - rol
    - terms
    type: object
  models.PublicRegistrationResponse:
    properties:
      account_type:
        $ref: '#/definitions/models.AccountType'
      event_date:
        type: string
      event_name:
        description: Event details
        type: string
      gebruiker_id:
        description: Extra info voor full accounts
        type: string
      has_app_access:
        type: boolean
      message:
        type: string
      participant_id:
        type: string
      registration_id:
        description: EventRegistration ID
        type: string
      success:
        type: boolean
    type: object
...
  models.UpgradeToFullAccountRequest:
    properties:
      email:
        type: string
      wachtwoord:
        minLength: 8
        type: string
    required:
    - email
    - wachtwoord
    type: object
  models.UpgradeToFullAccountResponse:
    properties:
      gebruiker_id:
        type: string
      has_app_access:
        type: boolean
      message:
        type: string
      participant_id:
        description: Participant info blijft hetzelfde maar heeft nu app toegang
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: info@dekoninklijkeloop.nl
    name: API Support
    url: http://www.dekoninklijkeloop.nl
  description: API voor de De Koninklijke Loop email service - inclusief registratie,
    deelnemersbeheer, CMS en meer
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DKL Email Service API
  version: "1.0"
paths:
...
  /api/events/active:
    get:
      consumes:
      - application/json
      description: Haalt het eerstvolgende actieve event op
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EventResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Actief Event ophalen
      tags:
      - Events
...
  /api/public/aanmelden:
    post:
      consumes:
      - application/json
      description: Registreert een new deelnemer met keuze tussen full account (met
        app toegang) of temporary registratie (alleen voor dit jaar)
      parameters:
      - description: Registratie gegevens
        in: body
        name: registration
        required: true
        schema:
          $ref: '#/definitions/models.PublicRegistrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PublicRegistrationResponse'
      summary: Registreer deelnemer voor evenement
      tags:
      - Public Registration
...
  /api/public/upgrade-to-full-account:
    post:
      consumes:
      - application/json
      description: Upgrade een tijdelijke registratie naar een volledig account met
        app toegang
      parameters:
      - description: Upgrade gegevens
        in: body
        name: upgrade
        required: true
        schema:
          $ref: '#/definitions/models.UpgradeToFullAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UpgradeToFullAccountResponse'
      summary: Upgrade naar volledig account
      tags:
      - Public Registration
...
securityDefinitions:
  BearerAuth:
    description: JWT Bearer token voor geauthenticeerde endpoints
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"